+++ tests/unit/Applet/appletFacadeMock.ts
@@ -1,7 +1,10 @@
 import * as sinon from 'sinon';
 
 export default {
-	getAppletName: sinon.fake(() => Promise.resolve('fakeAppletName')),
 	getAppletVersion: sinon.fake(() => Promise.resolve('fakeAppletVersion')),
 	getAppletUid: sinon.fake(() => Promise.resolve('fakeAppletUid')),
+	getApplet: sinon.fake(() => Promise.resolve({
+		name: 'fakeAppletName',
+		version: 'fakeAppletVersion',
+	})),
 };